slightly better result for actors train(10 per class) test(5 per class) 5/15 recognized

+++ dl4j-labs/src/main/java/global/skymind/training/convolution/objectdetection/transferlearning/tinyyolo/dataHelpers/XmlLabelProvider.java
@@ -33,7 +33,7 @@ public class XmlLabelProvider implements ImageObjectLabelProvider {
                 DocumentBuilder builder = factory.newDocumentBuilder();
                 Document doc = builder.parse(listOfFiles[i]);
                 NodeList objects = doc.getElementsByTagName("object");
-                int m = (int)objects.getLength();
+                int m = objects.getLength();
                 ArrayList<ImageObject> list = new ArrayList(m);
                 for(int j = 0; j < m; ++j) {
                     Element el = (Element)objects.item(j);
